ExtJS5 sends filter request a bit differently. It no longer sends a type information. So we have to rely on type autodetection to figure out which Filter instance to create. This was already implemented for String-, Boolean- and NumericFilter.
This commit adds support for the missing ListFilter. When value is of
type List a ListFilter is created. The ListFilter is now generic and is
no longer limited to String elements. For instance it's now possible to
send a list of numbers

This commit also adds a new field to the filter classes. rawConversion
which contains the raw value of the comparison (extjs4) / operator
(extjs5) property of the request. The conversion field still contains
the mapped Conversion enum value. But now the library no longer throws
an exception when the client sends an unknown operation/conversion
parameter. In this case rawConversion contains the sent value and the
property conversion is null.